+++ crates/objc2/src/topics/about_generated/CHANGELOG.md
@@ -27,6 +27,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - `CoreGraphics` / `objc2-core-graphics`.
   - `CoreFoundation` / `objc2-core-foundation`.
   - `CoreMedia` / `objc2-core-media`.
+  - `CoreMIDI` / `objc2-core-midi`.
   - `CoreText` / `objc2-core-text`.
   - `CoreVideo` / `objc2-core-video`.
   - `EventKitUI` / `objc2-event-kit-ui`.